Create Traffic-signs Recognition
In today’s world as the number of vehicles are increasing so are the road accidents and according to reports, many accident happens every minutes. This is caused due to many reasons such as poor enforcement of laws, carelessness etc. One of the reason is that people don’t recognize or follow traffic sign boards. So I made a traffic sign recognizer which can be used in smart cars to inform the driver of the vehicle about the traffic sign coming ahead and to follow it. This can reduce the road accidents.